+++ packages/unity-interface/BrowserInterface.ts
@@ -27,7 +27,7 @@ import {
 } from 'shared/types'
 import {
   getSceneWorkerBySceneID,
-  // setNewParcelScene,
+  setNewParcelScene,
   stopParcelSceneWorker,
   allScenesEvent,
   AllScenesEvents,
@@ -54,7 +54,6 @@ import { reportHotScenes } from 'shared/social/hotScenes'
 import { GIFProcessor } from 'gif-processor/processor'
 import { setVoiceChatRecording, setVoicePolicy, setVoiceVolume, toggleVoiceChatRecording } from 'shared/comms/actions'
 import { getERC20Balance } from 'shared/ethereum/EthereumService'
-// import { StatefulWorker } from 'shared/world/StatefulWorker'
 import { ensureFriendProfile } from 'shared/friends/ensureFriendProfile'
 import { reloadScene } from 'decentraland-loader/lifecycle/utils/reloadScene'
 import { wearablesRequest } from 'shared/catalogs/actions'
@@ -451,10 +450,13 @@ export class BrowserInterface {
         const options: StatefulWorkerOptions = {
           isEmpty: false
         }
-        options
-        throw new Error('Fix this circular dependency.')
 
-        // setNewParcelScene(sceneId, new StatefulWorker(parcelScene, options))
+        async function asyncLoad() {
+          const { StatefulWorker } = await import('shared/world/StatefulWorker')
+          setNewParcelScene(sceneId, new StatefulWorker(parcelScene, options))
+        }
+
+        asyncLoad().catch(defaultLogger.error)
         break
       }
       case 'StopStatefulMode': {
